fixed e2e tests
