+++ src/shared/PremoveSurvey/index.jsx
@@ -111,9 +111,7 @@ export function PreMoveSurveyEditablePanelify(DisplayComponent, EditComponent, e
     cancel = () => {
       this.props.reset();
       this.setIsEditable(false);
-      if (this.props.title === 'Pre-move survey') {
-        this.props.setEditPreMoveSurvey(false);
-      }
+      this.props.setEditPreMoveSurvey(false);
     };
     setIsEditable = isEditable => this.setState({ isEditable });
     render() {
